pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29.0",
    "fastapi>=0.109.0",
    "openai>=1.12.0",
    "openai-agents>=0.0.5",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "uvicorn>=0.27.0",
]
